+++ packages/syntax-codemods/src/run-js-codeshift.ts
@@ -2,9 +2,9 @@ import Runner from "jscodeshift/src/Runner";
 
 export default function runJscodeshift(
   transformerPath: string,
-  flags: Record<string, any>,
+  flags: Record<string, unknown>,
   files: string[],
-) {
+): ReturnType<typeof Runner.run> {
   return Runner.run(transformerPath, files, {
     ignorePattern: ["**/node_modules/**", "**/dist/**"],
     extensions: "tsx,ts,jsx,js",