+++ packages/syntax-codemods/src/transforms/__tests__/box-color-to-backgroundcolor.test.ts
@@ -49,6 +49,14 @@ describe("box-color-to-backgroundcolor", () => {
       },
     });
 
-    expect(format(String(output))).toEqual(format(outputCode));
+    expect(
+      format(String(output), {
+        parser: "babel",
+      }),
+    ).toStrictEqual(
+      format(outputCode, {
+        parser: "babel",
+      }),
+    );
   });
 });