+++ packages/syntax-core/src/Box/Box.test.tsx
@@ -1,6 +1,7 @@
 import { screen, render } from "@testing-library/react";
 import Box from "./Box";
 import { expect } from "vitest";
+import { createRef } from "react";
 
 describe("box", () => {
   it("renders successfully", () => {
@@ -11,7 +12,12 @@ describe("box", () => {
 
   it("passes trough certain props", () => {
     render(
-      <Box aria-labelledby="other-element" data-testid="testId" id="id">
+      <Box
+        aria-labelledby="other-element"
+        data-testid="testId"
+        id="id"
+        role="dialog"
+      >
         Test
       </Box>,
     );
@@ -19,6 +25,8 @@ describe("box", () => {
     const box = screen.getByTestId("testId");
     expect(box).toBeInTheDocument();
     expect(box.getAttribute("aria-labelledby")).toBe("other-element");
+    expect(box.getAttribute("data-testid")).toBe("testId");
+    expect(box.getAttribute("role")).toBe("dialog");
     expect(box.getAttribute("id")).toBe("id");
   });
 
@@ -97,5 +105,12 @@ describe("box", () => {
     const box = screen.getByTestId("testId");
     expect(box).toBeInTheDocument();
     expect(box.getAttribute("style")).toMatchInlineSnapshot('"opacity: 0.9;"');
+  });
+
+  it("allows for a ref to be set on Box", () => {
+    const ref = createRef<HTMLDivElement>();
+    render(<Box data-testid="test" ref={ref} />);
+    expect(ref.current instanceof HTMLDivElement).toBeTruthy();
+    expect(ref.current?.getAttribute("data-testid")).toStrictEqual("test");
   });
 });