+++ packages/syntax-core/src/Button/Button.stories.tsx
@@ -1,4 +1,4 @@
-import { StoryObj, Meta } from "@storybook/react";
+import { type StoryObj, type Meta } from "@storybook/react";
 import Button from "./Button";
 import FavoriteBorder from "@mui/icons-material/FavoriteBorder";