+++ packages/syntax-core/src/Button/Button.tsx
@@ -1,8 +1,8 @@
 import classNames from "classnames";
 import backgroundColor from "../colors//backgroundColor";
 import foregroundColor from "../colors/foregroundColor";
-import React, { ReactElement } from "react";
-import { Color, Size } from "../constants";
+import React, { forwardRef } from "react";
+import { type Color, type Size } from "../constants";
 import styles from "./Button.module.css";
 
 const textVariant = {
@@ -24,23 +24,11 @@ const iconSize = {
   ["lg"]: styles.lgIcon,
 };
 
-/**
- * Button is a clickable element that is used to perform an action.
- */
-const Button = ({
-  text,
-  loadingText,
-  color = "primary",
-  size = "md",
-  accessibilityLabel,
-  disabled = false,
-  loading = false,
-  fullWidth = false,
-  startIcon: StartIcon,
-  endIcon: EndIcon,
-  onClick,
-  tooltip,
-}: {
+type ButtonType = {
+  /**
+   * Test id for the button
+   */
+  "data-testid"?: string;
   /**
    * The text to be displayed inside the button
    */
@@ -103,61 +91,86 @@ const Button = ({
    * The tooltip to be displayed when the user hovers over the button
    */
   tooltip?: string;
-}): ReactElement => {
-  return (
-    <button
-      aria-label={accessibilityLabel}
-      type="button"
-      title={tooltip}
-      disabled={disabled || loading}
-      onClick={onClick}
-      className={classNames(
-        styles.button,
-        foregroundColor(color),
-        backgroundColor(color),
-        styles[size],
-        {
-          [styles.fullWidth]: fullWidth,
-          [styles.buttonGap]: size === "lg" || size === "md",
-          [styles.secondaryBorder]: color === "secondary",
-          [styles.secondaryDestructiveBorder]:
-            color === "destructive-secondary",
-        },
-      )}
-    >
-      {!loading && StartIcon && (
-        <StartIcon className={classNames(styles.icon, iconSize[size])} />
-      )}
-      {((loading && loadingText) || (!loading && text)) && (
-        <div className={styles.textContainer}>
-          {/* Replace with `Typography` once it lands in `syntax-core` */}
-          <div className={classNames(styles.buttonText, textVariant[size])}>
-            {loading ? loadingText : text}
-          </div>
-        </div>
-      )}
-      {!loading && EndIcon && (
-        <EndIcon className={classNames(styles.icon, iconSize[size])} />
-      )}
-      {loading && (
-        <svg
-          className={classNames(styles.loading, foregroundColor(color))}
-          viewBox="22 22 44 44"
-          width={loadingIconSize[size]}
-          height={loadingIconSize[size]}
-        >
-          <circle
-            className={styles.loadingCircle}
-            cx="44"
-            cy="44"
-            r="20.2"
-            fill="none"
-            strokeWidth="3.6"
-          />
-        </svg>
-      )}
-    </button>
-  );
 };
+
+const Button = forwardRef<HTMLButtonElement, ButtonType>(
+  (
+    {
+      "data-testid": dataTestId,
+      text,
+      loadingText,
+      color = "primary",
+      size = "md",
+      accessibilityLabel,
+      disabled = false,
+      loading = false,
+      fullWidth = false,
+      startIcon: StartIcon,
+      endIcon: EndIcon,
+      onClick,
+      tooltip,
+    },
+    ref,
+  ) => {
+    return (
+      <button
+        data-testid={dataTestId}
+        ref={ref}
+        aria-label={accessibilityLabel}
+        type="button"
+        title={tooltip}
+        disabled={disabled || loading}
+        onClick={onClick}
+        className={classNames(
+          styles.button,
+          foregroundColor(color),
+          backgroundColor(color),
+          styles[size],
+          {
+            [styles.fullWidth]: fullWidth,
+            [styles.buttonGap]: size === "lg" || size === "md",
+            [styles.secondaryBorder]: color === "secondary",
+            [styles.secondaryDestructiveBorder]:
+              color === "destructive-secondary",
+          },
+        )}
+      >
+        {!loading && StartIcon && (
+          <StartIcon className={classNames(styles.icon, iconSize[size])} />
+        )}
+        {((loading && loadingText) || (!loading && text)) && (
+          <div className={styles.textContainer}>
+            {/* Replace with `Typography` once it lands in `syntax-core` */}
+            <div className={classNames(styles.buttonText, textVariant[size])}>
+              {loading ? loadingText : text}
+            </div>
+          </div>
+        )}
+        {!loading && EndIcon && (
+          <EndIcon className={classNames(styles.icon, iconSize[size])} />
+        )}
+        {loading && (
+          <svg
+            className={classNames(styles.loading, foregroundColor(color))}
+            viewBox="22 22 44 44"
+            width={loadingIconSize[size]}
+            height={loadingIconSize[size]}
+          >
+            <circle
+              className={styles.loadingCircle}
+              cx="44"
+              cy="44"
+              r="20.2"
+              fill="none"
+              strokeWidth="3.6"
+            />
+          </svg>
+        )}
+      </button>
+    );
+  },
+);
+
+Button.displayName = "Button";
 
 export default Button;