+++ packages/syntax-core/src/Typography/Typography.stories.tsx
@@ -1,4 +1,4 @@
-import { StoryObj, Meta } from "@storybook/react";
+import { type StoryObj, type Meta } from "@storybook/react";
 import Typography from "./Typography";
 
 export default {
@@ -35,6 +35,9 @@ export default {
     },
     inline: {
       control: "boolean",
+    },
+    lineClamp: {
+      control: { type: "number", min: 0, max: 10, step: 1 },
     },
     size: {
       options: [100, 200, 300, 500, 600, 700, 800],