+++ packages/eslint-config-syntax/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @cambly/eslint-config-syntax
 
+## 1.3.1
+
+### Patch Changes
+
+- 28689b5: Internal: enable a few more TypeScript ESLint rules
+
+## 1.3.0
+
+### Minor Changes
+
+- e6c3e09: Update all dependencies to the latest versio
+
 ## 1.2.1
 
 ### Patch Changes