+++ packages/syntax-codemods/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @cambly/syntax-codemods
 
+## 0.5.1
+
+### Patch Changes
+
+- 28689b5: Internal: enable a few more TypeScript ESLint rules
+
+## 0.5.0
+
+### Minor Changes
+
+- e6c3e09: Update all dependencies to the latest versio
+
 ## 0.4.0
 
 ### Minor Changes